requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
pypdf>=3.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
